requirements.txt
anthropic>=0.40.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
pytest>=8.0.0
pytest-mock>=3.12.0
ruff>=0.5.0
mypy>=1.10.0
types-requests>=2.31.0
types-pyyaml>=6.0
google-cloud-aiplatform>=1.60.0
fastapi
uvicorn[standard]
jinja2
python-multipart
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
